mido_defconfig: Use a timer frequency of 100 Hz
The use of high CPU frequencies combined with a large number of cores makes latency decent with a 100 Hz scheduler tick. Reducing the tick rate from 300 Hz to 100 Hz improves throughput and significantly reduces the number of interrupts firing off per second, improving power consumption.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>

There are plenty of changes in the kernel for the performance side, balance it out with some battery improvements.